% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableILSA.R
\name{availableILSA}
\alias{availableILSA}
\title{Check available ILSA data}
\usage{
availableILSA(
  print = TRUE,
  FOR = c("download", "combine.students", "add.schools", "ILSAready")
)
}
\arguments{
\item{print}{a logical value indicating if results should be printed or not.}

\item{FOR}{a string indicating the availability of ILSA data for different 
purposes. Valid strings are \code{"download"}, \code{"combine.students"},
\code{add.schools}, and \code{"ILSAready"}.}
}
\value{
A list with the names of the ILSA and the available years.
}
\description{
Checks which 'SPSS' data from different International Large-Scale Assessments (ILSA).
are available.
}
\examples{

availableILSA(print = TRUE)

}
