% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistoNAs.R
\name{mistoNAs}
\alias{mistoNAs}
\title{Missing to NAs}
\usage{
mistoNAs(tibble)
}
\arguments{
\item{tibble}{a tibble object.}
}
\value{
A tibble.
}
\description{
Converts values labelled as missings to \code{NA}.
}
\examples{
# tibble generated by haven
input <- system.file("extdata/reds", package = "ILSAmerge")
tib <- do.call(rbind,justload(inputdir = input,population = "BCGV1"))

tib

mistoNAs(tib)

}
\keyword{tibblemanagement}
