% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uID.R
\name{uID}
\alias{uID}
\title{Create unique ID}
\usage{
uID(x, uID)
}
\arguments{
\item{x}{a data frame.}

\item{uID}{a string separated by ;.}
}
\value{
A character vector.
}
\description{
Creates a unique ID given a data frame and a string separated by ;.
}
