% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableILSA.R
\name{availableILSA}
\alias{availableILSA}
\title{Check available ILSA data}
\usage{
availableILSA(print = TRUE)
}
\arguments{
\item{print}{a logical value indicating if results should be printed or not.}
}
\value{
A list with the names of the ILSA and the available years.
}
\description{
Checks which 'SPSS' data from different International Large-Scale Assessments (ILSA).
are available.
}
\examples{

availableILSA(print = TRUE)

}
