% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILSAdownload.R
\name{ILSAdownload}
\alias{ILSAdownload}
\title{Downloads ILSA data}
\usage{
ILSAdownload(study, year, outputdir, unzip = FALSE)
}
\arguments{
\item{study}{a string indicating the name of the study. For available studies
check the description of this function.}

\item{year}{a numeric value indicating the year of the study.}

\item{outputdir}{the directory where the merged data will be saved.}

\item{unzip}{a logical value indicating if files should be unzipped.
Default is \code{FALSE}.}
}
\value{
Saves SPSS ILSA data locally.
}
\description{
Downloads SPSS data from different International Large-Scale Assessments (ILSA),
including: TIMSS, TIMSS Advanced, PIRLS, ICCS, ICILS, CIVED, REDS, RLII,
and SITES. Depending on the study, you will need to read and accept its
terms and conditions to proceed with the download.
}
\examples{
\dontrun{
# For example, to download TIMSS 1995 data:

# Path were files will be saved
output <- 'C:/'

# Downloading TIMSS 1995
ILSAdownload(study = "TIMSS", year = 1995, outputdir = output)

# Downloading TIMSS 1995 and unzipping files
ILSAdownload(study = "TIMSS", year = 1995, outputdir = output, unzip = TRUE)
}

}
