% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILSAmerge.R
\name{ILSAmerge}
\alias{ILSAmerge}
\title{Merge ILSA data}
\usage{
ILSAmerge(
  inputdir = getwd(),
  outputdir = getwd(),
  population = NULL,
  filetype = c("rds", "zsav", "sav"),
  MBlimit = NULL,
  MBlistlimit = 200,
  SPSSlimit = 50,
  quiet = FALSE
)
}
\arguments{
\item{inputdir}{a string indicating the path where ILSA 'SPSS' files are stored.}

\item{outputdir}{the directory where the merged data will be saved.}

\item{population}{a character vector indicating which files should be merged.
If \code{NULL} (the default), all files will be merged. For more information
on available populations, run \code{ILSAfile.info()} first.}

\item{filetype}{a string indicating the type of file to be saved, it can
be \code{"rds"}, \code{"zsav"}, or \code{"sav"}.}

\item{MBlimit}{a numerical value indicating the allowed limit of the combined
storage of the files of one type (see \code{ILSAfile.info()}).
For type files that go over the limit, files will not be merged in R,
but an 'SPSS' syntax will be produced via \code{spss.syntax()}.
If set to \code{NULL}, no limit will be used and all files will be merged
within R. If speed is a problem, we recommend
that this number should not be over \code{200} and merge the rest in 'SPSS'.
Beware that some ILSA will have files with different columns and this could
cause some 'SPSS' syntaxes to fail. If this happens, merge through \code{R}.}

\item{MBlistlimit}{a numerical value indicating the allowed limit of the
combined storage of the files of one type for merging through a list.
Values over the limit will be merged through a matrix, which will be slower
but uses less memory. Default is \code{200}.}

\item{SPSSlimit}{a numerical value indicating the limit of files per command
of 'SPSS', typically \code{50}.}

\item{quiet}{a logical value indicating if progress status should be
shown. Default is \code{FALSE}.}
}
\value{
Saves merged ILSA data or \code{.sps} syntax for merging ILSA data.
}
\description{
Merges 'SPSS' data from different International Large-Scale Assessments (ILSA).
This function has been tested to behave correctly for: 'TIMSS', 'TIMSS Advanced', 
'PIRLS', 'ICCS', 'ICILS', 'CIVED', 'REDS', 'RLII', and 'SITES' (2006).
}
\details{
For files merged within R it will also add country information where needed. 
Country information will be retrieved from 'GitHub' if possible. If not, it will
use the package internal data.
}
\examples{
# Path where raw 'SPSS' files are
input <- system.file("extdata/reds", package = "ILSAmerge")

# Path where merged files will be saved
dir.create(file.path(tempdir(),"ILSAmerge"))
output <- file.path(tempdir(),"ILSAmerge")

# Merging 'REDS' 2021, as .rds file
ILSAmerge(inputdir = input, outputdir = output, filetype = "rds", quiet = FALSE)

}
\keyword{ILSA}
