% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILSAready.R
\name{ILSAready}
\alias{ILSAready}
\title{Download and prepare ILSA data}
\usage{
ILSAready(
  study,
  year,
  outputdir = getwd(),
  filetype = c("rds", "zsav", "sav"),
  quiet = FALSE,
  agreeLicense = FALSE
)
}
\arguments{
\item{study}{a string indicating the name of the study. For available studies
check the description of this function.}

\item{year}{a numeric value indicating the year of the study.}

\item{outputdir}{the directory where the data will be saved.}

\item{filetype}{a string indicating the type of file to be saved, it can
be \code{"rds"}, \code{"zsav"}, or \code{"sav"}.}

\item{quiet}{a logical value indicating if status of progress should be
shown. 
Default is \code{FALSE}.}

\item{agreeLicense}{a logical value indicating if you agree with
the Disclaimer and License Agreement file from www.iea.nl. If \code{FALSE},
you will be prompted to agree with it or else data will not be downloaded.
Default is \code{FALSE}.}
}
\value{
Saves merged and renamed ILSA data.
}
\description{
Downloads ILSA data, merges it, combines students and add school information.
This function is a wrapper for \code{\link{ILSAdownload}}, \code{\link{ILSAmerge}},
\code{\link{ILSArename}}, \code{\link{combineStudents}}, and \code{\link{addSchools}}. 
To see which ILSA are available for this function use \code{\link{availableILSA}}.
}
\examples{

# Path were files will be saved
dir.create(file.path(tempdir(),"ILSAready"))
output <- file.path(tempdir(),"ILSAready")

# Downloading and preparing 'RLII' 1991 data
ILSAready("RLII",1991,outputdir = output, agreeLicense = TRUE)


}
