% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justload.R
\name{justload}
\alias{justload}
\title{Loading ILSA data into a list}
\usage{
justload(inputdir, population, justattributes = FALSE)
}
\arguments{
\item{inputdir}{a string indicating the path were ILSA 'SPSS' files are stored.}

\item{population}{a character value indicating which files should be merged. 
For more information on available populations, run \code{ILSAfile.info()} first.}

\item{justattributes}{a logical value indicating if 0 rows should be loaded.
This can be used when we just need to check column attributes. Default is 
\code{FALSE}.}
}
\value{
A list of tibbles.
}
\description{
Load 'SPSS' data from different International Large-Scale Assessments (ILSA),
including: 'TIMSS', 'TIMSS Advanced', 'PIRLS', 'ICCS', 'ICILS', 'CIVED', 'REDS', 'RLII',
and 'SITES' into a list.
}
\examples{
\donttest{
# For example, after downloading 'REDS' 2021 G4 data:

# Downloading 'REDS' 2021 and unzipping files
ILSAdownload(study = "REDS", year = 2021, outputdir = tempdir(), unzip = TRUE, agreeLicense = TRUE)

# Path were raw 'SPSS' files are
input <- file.path(tempdir(),"REDS2021_IDB_SPSS/Data")

# Load only attributes
justload(inputdir = input, population = "bcgv1", justattributes = TRUE)

# Load complete data
justload(inputdir = input, population = "bsgv1", justattributes = FALSE)
}

}
